Rename "template" to "strtemplate".
